requirements.txt
accelerate==1.3.0
av==14.1.0
certifi==2024.12.14
charset-normalizer==3.4.1
colpali_engine @ git+https://github.com/illuin-tech/colpali@59e94a92790b67bd60507608c3115a2e48f83a07
docx==0.2.4
filelock==3.17.0
fsspec==2024.12.0
GPUtil==1.4.0
huggingface-hub==0.28.0
idna==3.10
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.2.2
packaging==24.2
peft==0.14.0
pillow==11.1.0
psutil==6.1.1
PyMuPDF==1.25.2
python-docx==1.1.2
PyYAML==6.0.2
qwen-vl-utils==0.0.10
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
sympy==1.13.1
tokenizers==0.21.0
torch==2.6.0+cpu
torchaudio==2.6.0+cpu
torchvision==0.21.0+cpu
tqdm==4.67.1
transformers @ git+https://github.com/huggingface/transformers@e4227eb4d4cc1ac279865a12b52f957b818616f7
typing_extensions==4.12.2
urllib3==2.3.0
